requirements.txt
# Core dependencies
moviepy==1.0.3
Pillow>=10.0.0
numpy>=1.24.0
httpx>=0.25.0
pydantic>=2.0.0
PyYAML>=6.0
typer[all]>=0.9.0
rich>=13.0.0

# Audio processing
pydub>=0.25.0
soundfile>=0.12.0

# Optional: AI script generation
# openai>=1.0.0

# Development
pytest>=7.0.0
black>=23.0.0
ruff>=0.1.0
